Add CircleCiManualWebhookEvent
Since introducing a manual webhook to CircleCI to replace their default
webhook, we need to take a couple of things into consideration:

- the payload is not wrapped in a 'payload' document anymore;
- the 'outcome' field will always be 'running', since the manual webhook call is part of the build itself.

This means that we can't look at the 'outcome' field anymore. Instead, we check every action of every step of
the build to make sure it was successful, except for the last step, which is for the manual webhook call
itself.

Signed-off-by: Giuseppe Capizzi <gcapizzi@pivotal.io>