WIP for a locking worker
The idea is this worker could be a long running process running on every machine.

It uses a row lock to check that it is the only worker of that type operating with that database. This could be an effective design as the locking mechanism is built into the shared resource that causes the concurency issue.